add patrolCreature

+++ Assets/Script/Model/Weapon/Bludgeon.cs
@@ -4,5 +4,5 @@
     public float speed { get => -1f; }
     public float defence { get => 2f; }
     public float force {get => 5f; }
-    public float attackSpeed { get => -3f; }
+    public int attackSpeed { get => -3; }
 }
