+++ Assets/Script/Model/Weapon/IWeapon.cs
@@ -13,6 +13,6 @@
     float force { get; }
 
     /* Creature speed of striking */
-    float attackSpeed { get; }
+    int attackSpeed { get; }
 
 }
