+++ Assets/Script/Model/Weapon/Melee.cs
@@ -4,5 +4,5 @@
     public float speed { get => 0f; }
     public float defence { get => 0f; }
     public float force { get => 0f; }
-    public float attackSpeed { get => 0f; }
+    public int attackSpeed { get => 0; }
 }
